requirements.txt
# Python dependencies for data pipeline
kafka-python==2.0.2
pyspark==3.5.0